+++ Sources/VideoKit/VideoPlayerController.swift
@@ -13,9 +13,8 @@ import SwiftUI
 @MainActor
 public final class VideoPlayerController: AVPlayerViewController, VideoSessionManager {
 
-    /// TODO: Verify that this works, since it will crash if
-    /// we're not in an isolated context. If it doesn't work,
-    /// we need to call `destroyPlayer` from the outside.
+    /// TODO: Verify that this works, since it will crash if we're not in an isolated
+    /// context. If it doesn't, we must call `destroyPlayer` from the outside.
     deinit {
         MainActor.assumeIsolated {
             destroyPlayer()