+++ Sources/VideoKit/VideoSessionManager.swift
@@ -9,16 +9,14 @@
 import AVKit
 import SwiftUI
 
-/// This protocol can be implemented by any type that should
-/// be able to set up a video session.
+/// This protocol can be implemented by any type that can set up a video session.
 protocol VideoSessionManager {}
 
 extension VideoSessionManager {
 
     /// Set up the shared audio/video session for long video.
     ///
-    /// This function only has effect on iOS, macOS Catalyst,
-    /// tvOS, and visionOS.
+    /// This function only works on iOS, macOS Catalyst, tvOS, and visionOS.
     func setupVideoSession() throws {
         #if os(iOS) || targetEnvironment(macCatalyst) || os(visionOS)
         try AVAudioSession.sharedInstance()